modify loader logic
